@@ -23,18 +23,18 @@
     "issues": "https://github.com/chesio/bc-cache/issues"
   },
   "require": {
-    "php": "^7.3 || ^8.0",
+    "php": "^8.0",
     "composer/installers": "^1.0 || ^2.0"
   },
   "require-dev": {
     "brain/monkey": "^2.6",
-    "mockery/mockery": "^1.4",
+    "mockery/mockery": "^1.6",
     "php-parallel-lint/php-parallel-lint": "^1.3",
-    "php-stubs/wp-cli-stubs": "^2.7",
+    "php-stubs/wp-cli-stubs": "^2.8",
     "phpunit/phpunit": "^9.5",
     "squizlabs/php_codesniffer": "^3.3",
     "symfony/polyfill-php80": "^1.27",
-    "szepeviktor/phpstan-wordpress": "^1.0"
+    "szepeviktor/phpstan-wordpress": "^1.3"
   },
   "autoload-dev": {
     "psr-4": {
