@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace BlueChip\Cache;
 
 class Crawler
@@ -22,18 +24,8 @@ class Crawler
     private const DEFAULT_CACHE_WARM_UP_INVOCATION_DELAY = 10 * MINUTE_IN_SECONDS;
 
 
-    /**
-     * @var Feeder
-     */
-    private $cache_feeder;
-
-
-    /**
-     * @param Feeder $cache_feeder
-     */
-    public function __construct(Feeder $cache_feeder)
+    public function __construct(private Feeder $cache_feeder)
     {
-        $this->cache_feeder = $cache_feeder;
     }
 
 
