@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace BlueChip\Cache;
 
 class Viewer
@@ -24,25 +26,7 @@ class Viewer
      */
     private const START_WARM_UP_ACTION = 'start-warm-up';
 
-    /**
-     * @var Core
-     */
-    private $cache;
-
-    /**
-     * @var Crawler|null
-     */
-    private $cache_crawler;
-
-    /**
-     * @var Feeder|null
-     */
-    private $cache_feeder;
-
-    /**
-     * @var ListTable
-     */
-    private $list_table;
+    private ListTable $list_table;
 
 
     /**
@@ -50,11 +34,8 @@ class Viewer
      * @param Crawler|null $cache_crawler Null value signals that cache warm up is disabled.
      * @param Feeder|null $cache_feeder Null value signals that cache warm up is disabled.
      */
-    public function __construct(Core $cache, ?Crawler $cache_crawler, ?Feeder $cache_feeder)
+    public function __construct(private Core $cache, private ?Crawler $cache_crawler, private ?Feeder $cache_feeder)
     {
-        $this->cache = $cache;
-        $this->cache_crawler = $cache_crawler;
-        $this->cache_feeder = $cache_feeder;
     }
 
 
