@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace BlueChip\Cache;
 
 /**
@@ -10,27 +12,14 @@ namespace BlueChip\Cache;
 class XmlSitemapReader
 {
     /**
-     * @var string
-     */
-    private $robots_txt_url;
-
-    /**
-     * @var string
-     */
-    private $default_sitemap_url;
-
-
-    /**
      * @internal $robots_txt_url is parsed in order to determine URLs of available XML sitemaps.
      * If none are found, $default_sitemap_url is used as fallback.
      *
      * @param string $robots_txt_url
      * @param string $default_sitemap_url
      */
-    public function __construct(string $robots_txt_url, string $default_sitemap_url)
+    public function __construct(private string $robots_txt_url, private string $default_sitemap_url)
     {
-        $this->robots_txt_url = $robots_txt_url;
-        $this->default_sitemap_url = $default_sitemap_url;
     }
 
 
@@ -55,9 +44,7 @@ class XmlSitemapReader
         if (\preg_match_all('/^[Ss]itemap: ?(.+)$/m', $body, $matches)) {
             return \array_reduce(
                 $matches[1],
-                function (array $urls, string $sitemap_url): array {
-                    return \array_merge($urls, self::fetch($sitemap_url));
-                },
+                fn (array $urls, string $sitemap_url): array => \array_merge($urls, self::fetch($sitemap_url)),
                 []
             );
         } else {
@@ -89,19 +76,16 @@ class XmlSitemapReader
 
         try {
             $xml = new \SimpleXMLElement($body);
-        } catch (\Exception $e) {
-            throw new Exception("Could not parse {$url} as XML file - XML parser reports: " . $e->getMessage());
+        } catch (\Exception $exception) {
+            throw new Exception("Could not parse {$url} as XML file - XML parser reports: " . $exception->getMessage());
         }
 
-        switch ($xml->getName()) {
-            case 'sitemapindex':
-                // Ignore sitemap index if only sitemap is expected.
-                return $expect_sitemap ? [] : self::readUrlsFromSitemapIndex($xml);
-            case 'urlset':
-                return self::readUrlsFromSitemap($xml);
-            default:
-                return [];
-        }
+        return match ($xml->getName()) {
+            // Ignore sitemap index if only sitemap is expected.
+            'sitemapindex' => $expect_sitemap ? [] : self::readUrlsFromSitemapIndex($xml),
+            'urlset' => self::readUrlsFromSitemap($xml),
+            default => [],
+        };
     }
 
 
