@@ -7,7 +7,7 @@
 // Register autoload function
 \spl_autoload_register(function (string $class) {
     // Only autoload classes shipped with the plugin.
-    if (!\str_starts_with($class, 'BlueChip\\Cache')) {
+    if (!\str_starts_with($class, 'BlueChip\\Cache\\')) {
         return;
     }
 
